move Notifiable to OpenProject
