+++ app/helpers/settings_helper.rb
@@ -175,7 +175,7 @@ module SettingsHelper
     setting_label(setting, options) + wrap_field_outer(options, &block)
   end
 
-  # Renders a notification field for a Redmine::Notifiable option
+  # Renders a notification field for an OpenProject::Notifiable option
   def notification_field(notifiable, options = {})
     content_tag(:label, class: 'form--label-with-check-box' + (notifiable.parent.present? ? ' parent' : '')) do
       styled_check_box_tag('settings[notified_events][]',
